Split large file; dirEval is now based on energy/d², not 1/d²; occlusion is now calculated over distance, not number of blocks
