+++ src/main/java/com/sonicether/soundphysics/SPEfx.java
@@ -1,0 +1,184 @@
+package com.sonicether.soundphysics;
+
+import com.sonicether.soundphysics.config.ConfigManager;
+import com.sonicether.soundphysics.config.ReverbParams;
+import net.minecraft.util.math.Vec3d;
+import org.lwjgl.openal.AL10;
+import org.lwjgl.openal.AL11;
+import org.lwjgl.openal.ALC10;
+import org.lwjgl.openal.EXTEfx;
+
+import static com.sonicether.soundphysics.SPLog.*;
+
+public class SPEfx {
+
+    private static int auxFXSlot0;
+    private static int auxFXSlot1;
+    private static int auxFXSlot2;
+    private static int auxFXSlot3;
+    private static int reverb0;
+    private static int reverb1;
+    private static int reverb2;
+    private static int reverb3;
+    private static int directFilter0;
+    private static int sendFilter0;
+    private static int sendFilter1;
+    private static int sendFilter2;
+    private static int sendFilter3;
+
+    public static void syncReverbParams()
+    {
+        if (auxFXSlot0 != 0)
+        {
+            //Set the global reverb parameters and apply them to the effect and effectslot
+            setReverbParams(ReverbParams.getReverb0(), auxFXSlot0, reverb0);
+            setReverbParams(ReverbParams.getReverb1(), auxFXSlot1, reverb1);
+            setReverbParams(ReverbParams.getReverb2(), auxFXSlot2, reverb2);
+            setReverbParams(ReverbParams.getReverb3(), auxFXSlot3, reverb3);
+        }
+    }
+
+    static void setupEFX()
+    {
+        //Get current context and device
+        final long currentContext = ALC10.alcGetCurrentContext();
+        final long currentDevice = ALC10.alcGetContextsDevice(currentContext);
+        if (ALC10.alcIsExtensionPresent(currentDevice, "ALC_EXT_EFX")) {
+            log("EFX Extension recognized.");
+        } else {
+            logError("EFX Extension not found on current device. Aborting.");
+            return;
+        }
+
+        // Create auxiliary effect slots
+        auxFXSlot0 = EXTEfx.alGenAuxiliaryEffectSlots();
+        log("Aux slot " + auxFXSlot0 + " created");
+        EXTEfx.alAuxiliaryEffectSloti(auxFXSlot0, EXTEfx.AL_EFFECTSLOT_AUXILIARY_SEND_AUTO, AL10.AL_TRUE);
+
+        auxFXSlot1 = EXTEfx.alGenAuxiliaryEffectSlots();
+        log("Aux slot " + auxFXSlot1 + " created");
+        EXTEfx.alAuxiliaryEffectSloti(auxFXSlot1, EXTEfx.AL_EFFECTSLOT_AUXILIARY_SEND_AUTO, AL10.AL_TRUE);
+
+        auxFXSlot2 = EXTEfx.alGenAuxiliaryEffectSlots();
+        log("Aux slot " + auxFXSlot2 + " created");
+        EXTEfx.alAuxiliaryEffectSloti(auxFXSlot2, EXTEfx.AL_EFFECTSLOT_AUXILIARY_SEND_AUTO, AL10.AL_TRUE);
+
+        auxFXSlot3 = EXTEfx.alGenAuxiliaryEffectSlots();
+        log("Aux slot " + auxFXSlot3 + " created");
+        EXTEfx.alAuxiliaryEffectSloti(auxFXSlot3, EXTEfx.AL_EFFECTSLOT_AUXILIARY_SEND_AUTO, AL10.AL_TRUE);
+        checkErrorLog("Failed creating auxiliary effect slots!");
+
+        //Create effect objects
+        reverb0 = EXTEfx.alGenEffects();												//Create effect object
+        EXTEfx.alEffecti(reverb0, EXTEfx.AL_EFFECT_TYPE, EXTEfx.AL_EFFECT_EAXREVERB);		//Set effect object to be reverb
+        checkErrorLog("Failed creating reverb effect slot 0!");
+        reverb1 = EXTEfx.alGenEffects();												//Create effect object
+        EXTEfx.alEffecti(reverb1, EXTEfx.AL_EFFECT_TYPE, EXTEfx.AL_EFFECT_EAXREVERB);		//Set effect object to be reverb
+        checkErrorLog("Failed creating reverb effect slot 1!");
+        reverb2 = EXTEfx.alGenEffects();												//Create effect object
+        EXTEfx.alEffecti(reverb2, EXTEfx.AL_EFFECT_TYPE, EXTEfx.AL_EFFECT_EAXREVERB);		//Set effect object to be reverb
+        checkErrorLog("Failed creating reverb effect slot 2!");
+        reverb3 = EXTEfx.alGenEffects();												//Create effect object
+        EXTEfx.alEffecti(reverb3, EXTEfx.AL_EFFECT_TYPE, EXTEfx.AL_EFFECT_EAXREVERB);		//Set effect object to be reverb
+        checkErrorLog("Failed creating reverb effect slot 3!");
+
+        // Create filters
+        directFilter0 = EXTEfx.alGenFilters();
+        EXTEfx.alFilteri(directFilter0, EXTEfx.AL_FILTER_TYPE, EXTEfx.AL_FILTER_LOWPASS);
+        logGeneral("directFilter0: "+directFilter0);
+
+        sendFilter0 = EXTEfx.alGenFilters();
+        EXTEfx.alFilteri(sendFilter0, EXTEfx.AL_FILTER_TYPE, EXTEfx.AL_FILTER_LOWPASS);
+        logGeneral("filter0: "+sendFilter0);
+
+        sendFilter1 = EXTEfx.alGenFilters();
+        EXTEfx.alFilteri(sendFilter1, EXTEfx.AL_FILTER_TYPE, EXTEfx.AL_FILTER_LOWPASS);
+        logGeneral("filter1: "+sendFilter1);
+
+        sendFilter2 = EXTEfx.alGenFilters();
+        EXTEfx.alFilteri(sendFilter2, EXTEfx.AL_FILTER_TYPE, EXTEfx.AL_FILTER_LOWPASS);
+        logGeneral("filter2: "+sendFilter2);
+
+        sendFilter3 = EXTEfx.alGenFilters();
+        EXTEfx.alFilteri(sendFilter3, EXTEfx.AL_FILTER_TYPE, EXTEfx.AL_FILTER_LOWPASS);
+        logGeneral("filter3: "+sendFilter3);
+        checkErrorLog("Error creating lowpass filters!");
+
+        syncReverbParams();
+    }
+
+    protected static void setEnvironment(final int sourceID, final float sendGain0, final float sendGain1,
+                                       final float sendGain2, final float sendGain3, final float sendCutoff0, final float sendCutoff1,
+                                       final float sendCutoff2, final float sendCutoff3, final float directCutoff, final float directGain)
+    {
+        if (!ConfigManager.getConfig().enabled) return;
+        // Set reverb send filter values and set source to send to all reverb fx slots
+        EXTEfx.alFilterf(sendFilter0, EXTEfx.AL_LOWPASS_GAIN, sendGain0);
+        EXTEfx.alFilterf(sendFilter0, EXTEfx.AL_LOWPASS_GAINHF, sendCutoff0);
+        AL11.alSource3i(sourceID, EXTEfx.AL_AUXILIARY_SEND_FILTER, auxFXSlot0, 1, sendFilter0);
+        checkErrorLog("Set Environment filter0:");
+
+        EXTEfx.alFilterf(sendFilter1, EXTEfx.AL_LOWPASS_GAIN, sendGain1);
+        EXTEfx.alFilterf(sendFilter1, EXTEfx.AL_LOWPASS_GAINHF, sendCutoff1);
+        AL11.alSource3i(sourceID, EXTEfx.AL_AUXILIARY_SEND_FILTER, auxFXSlot1, 1, sendFilter1);
+        checkErrorLog("Set Environment filter1:");
+
+        EXTEfx.alFilterf(sendFilter2, EXTEfx.AL_LOWPASS_GAIN, sendGain2);
+        EXTEfx.alFilterf(sendFilter2, EXTEfx.AL_LOWPASS_GAINHF, sendCutoff2);
+        AL11.alSource3i(sourceID, EXTEfx.AL_AUXILIARY_SEND_FILTER, auxFXSlot2, 1, sendFilter2);
+        checkErrorLog("Set Environment filter2:");
+
+        EXTEfx.alFilterf(sendFilter3, EXTEfx.AL_LOWPASS_GAIN, sendGain3);
+        EXTEfx.alFilterf(sendFilter3, EXTEfx.AL_LOWPASS_GAINHF, sendCutoff3);
+        AL11.alSource3i(sourceID, EXTEfx.AL_AUXILIARY_SEND_FILTER, auxFXSlot3, 1, sendFilter3);
+        checkErrorLog("Set Environment filter3:");
+
+        EXTEfx.alFilterf(directFilter0, EXTEfx.AL_LOWPASS_GAIN, directGain);
+        EXTEfx.alFilterf(directFilter0, EXTEfx.AL_LOWPASS_GAINHF, directCutoff);
+        AL10.alSourcei(sourceID, EXTEfx.AL_DIRECT_FILTER, directFilter0);
+        checkErrorLog("Set Environment directFilter0:");
+
+        AL10.alSourcef(sourceID, EXTEfx.AL_AIR_ABSORPTION_FACTOR, (float) ConfigManager.getConfig().General.airAbsorption);
+        checkErrorLog("Set Environment airAbsorption:");
+    }
+
+    public static void setSoundPos(final int sourceID, final Vec3d pos) {
+        if (!ConfigManager.getConfig().enabled) return;
+        //System.out.println(pos);//TO DO
+        AL10.alSourcefv(sourceID, 4100, new float[]{(float)pos.x, (float)pos.y, (float)pos.z});
+    }
+
+    /*
+     * Applies the parameters in the enum ReverbParams to the main reverb effect.
+     */
+    protected static void setReverbParams(final ReverbParams r, final int auxFXSlot, final int reverbSlot)
+    {
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_DENSITY, r.density);		//Set default parameters
+        checkErrorLog("Error while assigning reverb density: " + r.density);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_DIFFUSION, r.diffusion);		//Set default parameters
+        checkErrorLog("Error while assigning reverb diffusion: " + r.diffusion);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_GAIN, r.gain);		//Set default parameters
+        checkErrorLog("Error while assigning reverb gain: " + r.gain);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_GAINHF, r.gainHF);		//Set default parameters
+        checkErrorLog("Error while assigning reverb gainHF: " + r.gainHF);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_DECAY_TIME, r.decayTime);		//Set default parameters
+        checkErrorLog("Error while assigning reverb decayTime: " + r.decayTime);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_DECAY_HFRATIO, r.decayHFRatio);		//Set default parameters
+        checkErrorLog("Error while assigning reverb decayHFRatio: " + r.decayHFRatio);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_REFLECTIONS_GAIN, r.reflectionsGain);		//Set default parameters
+        checkErrorLog("Error while assigning reverb reflectionsGain: " + r.reflectionsGain);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_LATE_REVERB_GAIN, r.lateReverbGain);		//Set default parameters
+        checkErrorLog("Error while assigning reverb lateReverbGain: " + r.lateReverbGain);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_LATE_REVERB_DELAY, r.lateReverbDelay);		//Set default parameters
+        checkErrorLog("Error while assigning reverb lateReverbDelay: " + r.lateReverbDelay);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_AIR_ABSORPTION_GAINHF, r.airAbsorptionGainHF);		//Set default parameters
+        checkErrorLog("Error while assigning reverb airAbsorptionGainHF: " + r.airAbsorptionGainHF);
+        EXTEfx.alEffectf(reverbSlot, EXTEfx.AL_EAXREVERB_ROOM_ROLLOFF_FACTOR, r.roomRolloffFactor);		//Set default parameters
+        checkErrorLog("Error while assigning reverb roomRolloffFactor: " + r.roomRolloffFactor);
+
+
+        //Attach updated effect object
+        EXTEfx.alAuxiliaryEffectSloti(auxFXSlot, EXTEfx.AL_EFFECTSLOT_EFFECT, reverbSlot);
+    }
+
+}
